+++ datacontract/imports/bigquery_importer.py
@@ -261,9 +261,9 @@ def map_type_from_bigquery(bigquery_type_str: str) -> str:
         "STRING": "string",
         "BYTES": "array",
         "INTEGER": "integer",
-        "INT64": "integer",
+        "INT64": "integer",  # for dbt-bigquery
         "FLOAT": "number",
-        "FLOAT64": "number",
+        "FLOAT64": "number",  # for dbt-bigquery
         "BOOLEAN": "boolean",
         "BOOL": "boolean",
         "TIMESTAMP": "timestamp",
@@ -274,6 +274,7 @@ def map_type_from_bigquery(bigquery_type_str: str) -> str:
         "BIGNUMERIC": "number",
         "GEOGRAPHY": "object",
         "JSON": "object",
+        "INTERVAL": "string",
     }
 
     if bigquery_type_str in type_mapping: